Conflicts:
	eZ/Publish/Core/Persistence/Legacy/Content/Gateway/EzcDatabase/QueryBuilder.php
	eZ/Publish/Core/Persistence/Legacy/Content/Mapper.php
	eZ/Publish/Core/Persistence/Legacy/Tests/Content/ContentHandlerTest.php
	eZ/Publish/Core/Persistence/Legacy/Tests/Content/Gateway/EzcDatabaseTest.php
	eZ/Publish/Core/Persistence/Legacy/Tests/Content/MapperTest.php

+++ eZ/Publish/Core/Persistence/Legacy/Content/Language/Handler.php
@@ -11,7 +11,8 @@ namespace eZ\Publish\Core\Persistence\Legacy\Content\Language;
 use eZ\Publish\SPI\Persistence\Content\Language,
     eZ\Publish\SPI\Persistence\Content\Language\Handler as BaseLanguageHandler,
     eZ\Publish\SPI\Persistence\Content\Language\CreateStruct,
-    eZ\Publish\Core\Base\Exceptions\NotFoundException;
+    eZ\Publish\Core\Base\Exceptions\NotFoundException,
+    eZ\Publish\Core\Base\Exceptions\Logic;
 
 /**
  * Language Handler
@@ -78,8 +79,9 @@ class Handler implements BaseLanguageHandler
      */
     public function load( $id )
     {
-        $rows = $this->languageGateway->loadLanguageData( $id );
-        $languages = $this->languageMapper->extractLanguagesFromRows( $rows );
+        $languages = $this->languageMapper->extractLanguagesFromRows(
+            $this->languageGateway->loadLanguageData( $id )
+        );
 
         if ( count( $languages ) < 1 )
         {
@@ -97,8 +99,9 @@ class Handler implements BaseLanguageHandler
      */
     public function loadByLanguageCode( $languageCode )
     {
-        $rows = $this->languageGateway->loadLanguageDataByLanguageCode( $languageCode );
-        $languages = $this->languageMapper->extractLanguagesFromRows( $rows );
+        $languages = $this->languageMapper->extractLanguagesFromRows(
+            $this->languageGateway->loadLanguageDataByLanguageCode( $languageCode )
+        );
 
         if ( count( $languages ) < 1 )
         {
@@ -114,20 +117,25 @@ class Handler implements BaseLanguageHandler
      */
     public function loadAll()
     {
-        $rows = $this->languageGateway->loadAllLanguagesData();
-        return $this->languageMapper->extractLanguagesFromRows( $rows );
+        return $this->languageMapper->extractLanguagesFromRows(
+            $this->languageGateway->loadAllLanguagesData()
+        );
     }
 
     /**
      * Delete a language
      *
-     * @todo Might throw an exception if the language is still associated with
-     *       some content / types / (...) ?
-     *
      * @param mixed $id
+     *
+     * @throws \eZ\Publish\Core\Base\Exceptions\Logic If language could not be deleted
      */
     public function delete( $id )
     {
+        if ( !$this->languageGateway->canDeleteLanguage( $id ) )
+        {
+            throw new Logic( "Deleting language", "some content still references that language and therefore it can't be deleted" );
+        }
+
         $this->languageGateway->deleteLanguage( $id );
     }
 }
